feat(hodo-list): Move abilities from modal to page
