+++ apps/hodo-list/src/atoms/abilities-atom.ts
@@ -1,0 +1,5 @@
+import { atomWithStorage } from "jotai/utils";
+
+import type { Ability } from "@/types/ability.type";
+
+export const abilitiesAtom = atomWithStorage<Ability[]>("abilities", []);
