+++ apps/hodo-list/src/types/ability.type.ts
@@ -1,6 +1,6 @@
 export type Ability = {
   id: string;
-  title: string;
+  name: string;
   createdAt: number;
   value: number;
 };
